ml-service/requirements.txt
Flask==2.3.2
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
pymongo==4.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
flask-cors==4.0.0
joblib==1.3.1
pytest==7.4.0
pytest-flask==1.2.0